Rename bokeh.core.property.{undefined->singletons}
