+++ bokeh/core/property/singletons.py
